cmdline: factor out a cmdline handler to make option parsing generic
The cmdline parser can now be used for parsing commandlines other than
just what we expect for patch. This is done by factoring out a class
that has a virtual process_option function. Inheriting command line
handlers can implement this to interpret the given option in the
application specific method.
